Add generic Unicode converter and move BRF unicode converter to a separate file
